+++ src/components/Detail/comments/ReCommentLikes.tsx
@@ -7,7 +7,7 @@ import styled from 'styled-components';
 import supabase from 'src/lib/supabaseClient';
 import { IconLiked, IconUnLiked } from 'src/components/icons';
 import { useLocation } from 'react-router';
-import { NON_MEMBER } from 'src/function/alertMessage';
+import { NON_MEMBER } from 'src/utility/alertMessage';
 import { styleFont } from 'src/styles/styleFont';
 
 interface Props {
