+++ src/shared/PrivateRoute.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Navigate, Outlet, useLocation, useNavigate } from 'react-router';
-import { NON_MEMBER } from 'src/function/alertMessage';
+import { NON_MEMBER } from 'src/utility/alertMessage';
 
 export const PrivateRoute = () => {
   const navigate = useNavigate();
